// apps/frontend/packages/api/base/codegen/types/PlanRemark.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { PlanRemarkStatus2 } from "./PlanRemarkStatus2";

export interface PlanRemark {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type string
    */
    humanId: string;
    /**
     * @type integer
    */
    planId: number;
    /**
     * @description null — замечание от ДТОиР
     * @type integer,null
    */
    cfoId?: number | null;
    /**
     * @type integer
    */
    authorId: number;
    /**
     * @type string, date-time
    */
    createdAt: string;
    /**
     * @type integer,null
    */
    relatedSlotId?: number | null;
    /**
     * @type integer,null
    */
    fileVersionId?: number | null;
    /**
     * @type string
    */
    sheetName: string;
    /**
     * @type string
    */
    rowRef: string;
    /**
     * @type string
    */
    cellRef: string;
    /**
     * @type string
    */
    description: string;
    /**
     * @type string
    */
    requiredAction: string;
    /**
     * @type string
    */
    status: PlanRemarkStatus2;
    /**
     * @type integer,null
    */
    closedById?: number | null;
    /**
     * @type string,null, date-time
    */
    closedAt?: string | null;
    /**
     * @description Код ЦФО либо «ДТОиР»
     * @type string
    */
    issuerLabel: string;
}